requirements.txt
tensorflow==2.16.1
tf_keras==2.16.0
transformers==4.42.1
deepgram-sdk==2.12.0
flask==3.0.3
flask-socketio